+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/RxDocumentClientImpl.java
@@ -4661,7 +4661,7 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
                 Mono<Void> queryValidationMono;
                 if (customQuery != null) {
                     queryValidationMono = validateCustomQueryForReadManyByPartitionKeys(
-                        customQuery, resourceLink, state.getQueryOptions());
+                        customQuery, resourceLink, state.getQueryOptions(), collection);
                 } else {
                     queryValidationMono = Mono.empty();
                 }
@@ -5110,7 +5110,8 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
     private Mono<Void> validateCustomQueryForReadManyByPartitionKeys(
         SqlQuerySpec customQuery,
         String resourceLink,
-        CosmosQueryRequestOptions queryRequestOptions) {
+        CosmosQueryRequestOptions queryRequestOptions,
+        DocumentCollection collection) {
 
         IDocumentQueryClient queryClient = documentQueryClientImpl(
             RxDocumentClientImpl.this, getOperationContextAndListenerTuple(queryRequestOptions));
@@ -5122,6 +5123,7 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
                 customQuery,
                 resourceLink,
                 queryRequestOptions,
+                collection,
                 Configs.isQueryPlanCachingEnabled(),
                 this.getQueryPlanCache())
             .doOnNext(RxDocumentClientImpl::validateQueryPlanForReadManyByPartitionKeys)
@@ -5637,6 +5639,11 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
             @Override
             public GlobalPartitionEndpointManagerForPerPartitionCircuitBreaker getGlobalPartitionEndpointManagerForCircuitBreaker() {
                 return RxDocumentClientImpl.this.globalPartitionEndpointManagerForPerPartitionCircuitBreaker;
+            }
+
+            @Override
+            public boolean useThinClient(RxDocumentServiceRequest request) {
+                return RxDocumentClientImpl.this.useThinClientStoreModel(request);
             }
         };
     }
@@ -7424,7 +7431,8 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
             resourceType == ResourceType.ClientEncryptionKey ||
             resourceType.isScript() && operationType != OperationType.ExecuteJavaScript ||
             resourceType == ResourceType.PartitionKeyRange ||
-            resourceType == ResourceType.PartitionKey && operationType == OperationType.Delete) {
+            resourceType == ResourceType.PartitionKey && operationType == OperationType.Delete ||
+            operationType == OperationType.QueryPlan) {
             return this.gatewayProxy;
         }
 
@@ -7462,7 +7470,7 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
             if ((operationType == OperationType.Query ||
                 operationType == OperationType.SqlQuery ||
                 operationType == OperationType.ReadFeed) &&
-                    Utils.isCollectionChild(request.getResourceType())) {
+                Utils.isCollectionChild(request.getResourceType())) {
                 // Go to gateway only when partition key range and partition key are not set. This should be very rare
                 if (request.getPartitionKeyRangeIdentity() == null &&
                         request.getHeaders().get(HttpConstants.HttpHeaders.PARTITION_KEY) == null) {
@@ -9019,7 +9027,8 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
     private boolean useThinClientStoreModel(RxDocumentServiceRequest request) {
         if (!useThinClient
             || !this.globalEndpointManager.hasThinClientReadLocations()
-            || request.getResourceType() != ResourceType.Document) {
+            || (request.getResourceType() != ResourceType.Document
+                && !request.isExecuteStoredProcedureBasedRequest())) {
 
             return false;
         }
@@ -9029,7 +9038,10 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
         return operationType.isPointOperation()
                     || operationType == OperationType.Query
                     || operationType == OperationType.Batch
-                    || request.isChangeFeedRequest() && !request.isAllVersionsAndDeletesChangeFeedMode();
+                    || (request.isChangeFeedRequest()
+                        && !request.isAllVersionsAndDeletesChangeFeedMode())
+                    || request.isExecuteStoredProcedureBasedRequest()
+                    || operationType == OperationType.QueryPlan;
     }
 
     private DocumentClientRetryPolicy getRetryPolicyForPointOperation(